requirements/test.txt
coverage==7.6.1
pytest==8.3.2
pytest-cov==6.0.0
pytest-mock==3.14.0
pytest-xdist==3.6.1
pytest-xprocess==1.0.2
requests==2.32.3
